+++ packages/embedded-document-artefact-loader/src/source-map-utils.ts
@@ -4,17 +4,17 @@ import {
   RawSourceMap,
 } from "source-map-js";
 
+function toRawSourceMap(sourceMap: string | RawSourceMap): RawSourceMap {
+  return typeof sourceMap === "string" ? JSON.parse(sourceMap) : sourceMap;
+}
+
 export function applySourceMap(
   sourcePath: string,
   previousSourceMap: string | RawSourceMap,
-  nextSourceMap: string,
+  nextSourceMap: string | RawSourceMap,
 ): SourceMapGenerator {
-  const prev = new SourceMapConsumer(
-    typeof previousSourceMap === "string"
-      ? JSON.parse(previousSourceMap)
-      : previousSourceMap,
-  );
-  const next = new SourceMapConsumer(JSON.parse(nextSourceMap));
+  const prev = new SourceMapConsumer(toRawSourceMap(previousSourceMap));
+  const next = new SourceMapConsumer(toRawSourceMap(nextSourceMap));
   const pipeline = SourceMapGenerator.fromSourceMap(next);
   pipeline.applySourceMap(prev, sourcePath);
   return pipeline;