+++ packages/graphql-codegen-typescript-operations/src/config.ts
@@ -120,5 +120,4 @@ export interface TypeScriptDocumentsPluginConfig extends RawDocumentsConfig {
    */
   addOperationExport?: boolean;
   baseTypesPath?: string;
-  legacyEnumsCompatibility?: boolean;
 }