+++ packages/rempl-apollo-devtools/src/subscriber/apollo-operations-tracker/affected-queries/affected-queries-grid-renderer-styles.ts
@@ -1,7 +1,0 @@
-import { makeStyles } from "@fluentui/react-components";
-
-export const useStyles = makeStyles({
-  root: {
-    minHeight: 0,
-  },
-});