+++ packages/ts-codegen/src/__tests__/index.test.ts
@@ -573,7 +573,10 @@ describe(generateTS, () => {
         }
       `);
       expect(enums).toMatchInlineSnapshot(`
-        "export type Rank = "User" | "Admin";
+        "export enum Rank {
+            User = "User",
+            Admin = "Admin"
+        }
         "
       `);
       expect(inputs).toMatchInlineSnapshot(`
@@ -702,7 +705,11 @@ describe(generateTS, () => {
         }
       `);
       expect(enums).toMatchInlineSnapshot(`
-        "export type PresenceAvailability = "Available" | "Away" | "Offline";
+        "export enum PresenceAvailability {
+            Available = "Available",
+            Away = "Away",
+            Offline = "Offline"
+        }
         "
       `);
       expect(inputs).toMatchInlineSnapshot(`undefined`);
@@ -782,7 +789,7 @@ describe(generateTS, () => {
         export interface User extends BaseModel {
             readonly __typename?: "User";
             readonly id: string;
-            readonly availability: Enums.PresenceAvailability | \`\${Enums.PresenceAvailability}\`;
+            readonly availability: Enums.PresenceAvailability;
         }
         "
       `);
@@ -1214,7 +1221,10 @@ describe(generateTS, () => {
       }
     `);
     expect(enums).toMatchInlineSnapshot(`
-      "export type UserType = "Admin" | "User";
+      "export enum UserType {
+          Admin = "Admin",
+          User = "User"
+      }
       "
     `);
     expect(inputs).toMatchInlineSnapshot(`
@@ -1702,7 +1712,10 @@ describe(generateTS, () => {
       { legacyCompat: true },
     );
     expect(enums).toMatchInlineSnapshot(`
-      "export type Type = "type1" | "type2";
+      "export enum Type {
+          type1 = "type1",
+          type2 = "type2"
+      }
       "
     `);
     expect(inputs).toMatchInlineSnapshot(`undefined`);
@@ -1785,6 +1798,170 @@ describe(generateTS, () => {
       }
       "
     `);
+  });
+
+  it("generateTS with string unions instead of enums", () => {
+    const { models, resolvers, legacyTypes, enums, inputs } = runGenerateTest(
+      graphql`
+        interface Node {
+          id: ID!
+        }
+
+        enum Type {
+          type1
+          type2
+        }
+
+        type User implements Node {
+          id: ID!
+          userType: Type!
+        }
+
+        extend type Query {
+          user(id: ID!): User!
+        }
+      `,
+      { useStringUnionsInsteadOfEnums: true },
+    );
+    expect(enums).toMatchInlineSnapshot(`
+      "export type Type = "type1" | "type2";
+      "
+    `);
+    expect(inputs).toMatchInlineSnapshot(`undefined`);
+    expect(models).toMatchInlineSnapshot(`
+      "import * as Enums from "./enums.interface";
+      export * from "./enums.interface";
+      // Base type for all models. Enables automatic resolution of abstract GraphQL types (interfaces, unions)
+      export interface BaseModel {
+          readonly __typename?: string;
+      }
+      export interface Node extends BaseModel {
+          readonly __typename?: string;
+      }
+      export interface User extends BaseModel, Node {
+          readonly __typename?: "User";
+          readonly id: string;
+          readonly userType: Enums.Type;
+      }
+      "
+    `);
+    expect(resolvers).toMatchInlineSnapshot(`
+      "import type { PromiseOrValue } from "@graphitation/supermassive";
+      import type { ResolveInfo } from "@graphitation/supermassive";
+      import * as Models from "./models.interface";
+      export declare namespace Node {
+          export interface Resolvers {
+              readonly __resolveType?: __resolveType;
+          }
+          export type __resolveType = (parent: unknown, context: unknown, info: ResolveInfo) => PromiseOrValue<string | null>;
+      }
+      export declare namespace User {
+          export interface Resolvers {
+              readonly id?: id;
+              readonly userType?: userType;
+          }
+          export type id = (model: Models.User, args: {}, context: unknown, info: ResolveInfo) => PromiseOrValue<string>;
+          export type userType = (model: Models.User, args: {}, context: unknown, info: ResolveInfo) => PromiseOrValue<Models.Type>;
+      }
+      export declare namespace Query {
+          export interface Resolvers {
+              readonly user?: user;
+          }
+          export type user = (model: unknown, args: {
+              readonly id: string;
+          }, context: unknown, info: ResolveInfo) => PromiseOrValue<Models.User>;
+      }
+      "
+    `);
+    expect(legacyTypes).toMatchInlineSnapshot(`undefined`);
+  });
+
+  it("generateTS with string unions instead of enums, but only Enums specified in enumNamesToMigrate will be migrated", () => {
+    const { models, resolvers, legacyTypes, enums, inputs } = runGenerateTest(
+      graphql`
+        interface Node {
+          id: ID!
+        }
+
+        enum Type {
+          type1
+          type2
+        }
+
+        enum TypeToMigrate {
+          type1
+          type2
+        }
+
+        type User implements Node {
+          id: ID!
+          userType: Type!
+        }
+
+        extend type Query {
+          user(id: ID!): User!
+        }
+      `,
+      {
+        useStringUnionsInsteadOfEnums: true,
+        enumNamesToMigrate: ["TypeToMigrate"],
+      },
+    );
+    expect(enums).toMatchInlineSnapshot(`
+      "export enum Type {
+          type1 = "type1",
+          type2 = "type2"
+      }
+      export type TypeToMigrate = "type1" | "type2";
+      "
+    `);
+    expect(inputs).toMatchInlineSnapshot(`undefined`);
+    expect(models).toMatchInlineSnapshot(`
+      "import * as Enums from "./enums.interface";
+      export * from "./enums.interface";
+      // Base type for all models. Enables automatic resolution of abstract GraphQL types (interfaces, unions)
+      export interface BaseModel {
+          readonly __typename?: string;
+      }
+      export interface Node extends BaseModel {
+          readonly __typename?: string;
+      }
+      export interface User extends BaseModel, Node {
+          readonly __typename?: "User";
+          readonly id: string;
+          readonly userType: Enums.Type;
+      }
+      "
+    `);
+    expect(resolvers).toMatchInlineSnapshot(`
+      "import type { PromiseOrValue } from "@graphitation/supermassive";
+      import type { ResolveInfo } from "@graphitation/supermassive";
+      import * as Models from "./models.interface";
+      export declare namespace Node {
+          export interface Resolvers {
+              readonly __resolveType?: __resolveType;
+          }
+          export type __resolveType = (parent: unknown, context: unknown, info: ResolveInfo) => PromiseOrValue<string | null>;
+      }
+      export declare namespace User {
+          export interface Resolvers {
+              readonly id?: id;
+              readonly userType?: userType;
+          }
+          export type id = (model: Models.User, args: {}, context: unknown, info: ResolveInfo) => PromiseOrValue<string>;
+          export type userType = (model: Models.User, args: {}, context: unknown, info: ResolveInfo) => PromiseOrValue<Models.Type>;
+      }
+      export declare namespace Query {
+          export interface Resolvers {
+              readonly user?: user;
+          }
+          export type user = (model: unknown, args: {
+              readonly id: string;
+          }, context: unknown, info: ResolveInfo) => PromiseOrValue<Models.User>;
+      }
+      "
+    `);
+    expect(legacyTypes).toMatchInlineSnapshot(`undefined`);
   });
 
   it("legacy interfaces", () => {
@@ -2006,6 +2183,8 @@ function runGenerateTest(
     enumsImport?: string;
     legacyNoModelsForObjects?: boolean;
     legacyEnumsCompatibility?: boolean;
+    useStringUnionsInsteadOfEnums?: boolean;
+    enumNamesToMigrate?: string[];
     modelScope?: string;
   } = {},
 ): {
@@ -2017,6 +2196,8 @@ function runGenerateTest(
   legacyResolvers?: string;
   legacyNoModelsForObjects?: boolean;
   legacyEnumsCompatibility?: boolean;
+  useStringUnionsInsteadOfEnums?: boolean;
+  enumNamesToMigrate?: string[];
   modelScope?: string;
 } {
   const fullOptions: {
@@ -2027,6 +2208,8 @@ function runGenerateTest(
     legacyCompat?: boolean;
     legacyEnumsCompatibility?: boolean;
     legacyNoModelsForObjects?: boolean;
+    useStringUnionsInsteadOfEnums?: boolean;
+    enumNamesToMigrate?: string[];
   } = {
     outputPath: "__generated__",
     documentPath: "./typedef.graphql",