+++ examples/supermassive-todomvc/CHANGELOG.md
@@ -1,8 +1,54 @@
 # Change Log - supermassive-todomvc
 
-This log was last generated on Thu, 28 Sep 2023 12:04:15 GMT and should not be manually modified.
+This log was last generated on Mon, 16 Oct 2023 10:17:21 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.0.1-0
+
+Mon, 16 Oct 2023 10:17:21 GMT
+
+### Patches
+
+- Bump @graphitation/apollo-react-relay-duct-tape to v1.0.19
+- Bump @graphitation/apollo-react-relay-duct-tape-compiler to v1.3.5
+- Bump @graphitation/supermassive to v3.2.3
+
+## 1.0.1-0
+
+Mon, 16 Oct 2023 09:30:34 GMT
+
+### Patches
+
+- Bump @graphitation/apollo-react-relay-duct-tape to v1.0.18
+- Bump @graphitation/apollo-react-relay-duct-tape-compiler to v1.3.4
+- Bump @graphitation/supermassive to v3.2.2
+
+## 1.0.1-0
+
+Fri, 13 Oct 2023 08:43:43 GMT
+
+### Patches
+
+- Bump @graphitation/apollo-react-relay-duct-tape to v1.0.17
+
+## 1.0.1-0
+
+Fri, 13 Oct 2023 05:52:19 GMT
+
+### Patches
+
+- Bump @graphitation/apollo-react-relay-duct-tape to v1.0.16
+
+## 1.0.1-0
+
+Sun, 08 Oct 2023 15:21:54 GMT
+
+### Patches
+
+- Bump @graphitation/apollo-react-relay-duct-tape to v1.0.15
+- Bump @graphitation/apollo-react-relay-duct-tape-compiler to v1.3.3
+- Bump @graphitation/supermassive to v3.2.1
 
 ## 1.0.1-0