+++ packages/apollo-mock-client/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @graphitation/apollo-mock-client
 
-This log was last generated on Thu, 21 Sep 2023 06:28:45 GMT and should not be manually modified.
+This log was last generated on Fri, 13 Oct 2023 08:43:43 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.11.4
+
+Fri, 13 Oct 2023 08:43:43 GMT
+
+### Patches
+
+- Bump @graphitation/graphql-js-operation-payload-generator to v0.12.5
+
+## 0.11.3
+
+Fri, 13 Oct 2023 05:52:19 GMT
+
+### Patches
+
+- Bump @graphitation/graphql-js-operation-payload-generator to v0.12.4
 
 ## 0.11.2